Split case evidence pack review section
