+++ apps/operator-ui/src/app/operatorConsolePages/caseDetailSurfaces.tsx
@@ -42,6 +42,7 @@ import {
   useOperatorRecord,
   ValueList,
 } from "./shared";
+import { EvidencePackReviewSection } from "./caseDetailEvidencePackSection";
 
 export function CaseDetailPageBody({
   caseId,
@@ -278,107 +279,7 @@ export function CaseDetailPageBody({
         )}
       </SectionCard>
 
-      <SectionCard
-        subtitle="Evidence packs show backend-reviewed freshness, custody, confidence, provenance, conflict, uncertainty, and source posture as subordinate review context only."
-        title="Evidence pack review"
-      >
-        {evidencePacks.length > 0 ? (
-          <Table aria-label="Linked evidence packs" size="small">
-            <TableHead>
-              <TableRow>
-                <TableCell>Evidence pack</TableCell>
-                <TableCell>Source</TableCell>
-                <TableCell>Freshness</TableCell>
-                <TableCell>Confidence</TableCell>
-                <TableCell>Conflict</TableCell>
-                <TableCell>Uncertainty</TableCell>
-                <TableCell>Source state</TableCell>
-                <TableCell>Custody</TableCell>
-                <TableCell>Provenance</TableCell>
-                <TableCell>Boundary</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {evidencePacks.map((pack, index) => {
-                const freshnessState = asString(pack.freshness_state);
-                const conflictState = asString(pack.conflict_state);
-                const sourceState = asString(pack.source_state);
-                const custody = asRecord(pack.custody);
-                const provenance = asRecord(pack.provenance);
-                const packId = asString(pack.evidence_request_id);
-
-                return (
-                  <TableRow key={`${packId ?? "evidence-pack"}-${index}`}>
-                    <TableCell>{formatValue(pack.evidence_request_id)}</TableCell>
-                    <TableCell>{formatValue(pack.source_id)}</TableCell>
-                    <TableCell>
-                      <Chip
-                        color={
-                          freshnessState === "stale_review_required"
-                            ? "warning"
-                            : statusTone(freshnessState)
-                        }
-                        label={freshnessState ?? "unknown"}
-                        size="small"
-                        variant="outlined"
-                      />
-                    </TableCell>
-                    <TableCell>{formatValue(pack.confidence_state)}</TableCell>
-                    <TableCell>
-                      <Chip
-                        color={evidencePackConflictTone(conflictState)}
-                        label={conflictState ?? "unknown"}
-                        size="small"
-                        variant="outlined"
-                      />
-                    </TableCell>
-                    <TableCell>{formatValue(pack.uncertainty_label)}</TableCell>
-                    <TableCell>
-                      <Chip
-                        color={evidencePackSourceTone(sourceState)}
-                        label={sourceState ?? "unknown"}
-                        size="small"
-                        variant="outlined"
-                      />
-                    </TableCell>
-                    <TableCell>
-                      <Stack spacing={0.5}>
-                        <Typography variant="body2">{formatValue(pack.custody_state)}</Typography>
-                        <Typography color="text.secondary" variant="caption">
-                          {formatValue(
-                            custody?.aegisops_evidence_record_id ??
-                              custody?.enrichment_request_id ??
-                              custody?.collection_timestamp,
-                          )}
-                        </Typography>
-                      </Stack>
-                    </TableCell>
-                    <TableCell>
-                      <Stack spacing={0.5}>
-                        <Typography variant="body2">{formatValue(pack.provenance_state)}</Typography>
-                        <Typography color="text.secondary" variant="caption">
-                          {formatValue(
-                            provenance?.custody_reference ??
-                              provenance?.request_binding ??
-                              provenance?.source_id,
-                          )}
-                        </Typography>
-                      </Stack>
-                    </TableCell>
-                    <TableCell>
-                      <Typography color="text.secondary" variant="body2">
-                        Subordinate evidence context only
-                      </Typography>
-                    </TableCell>
-                  </TableRow>
-                );
-              })}
-            </TableBody>
-          </Table>
-        ) : (
-          <EmptyState message="No linked evidence packs were returned for this case." />
-        )}
-      </SectionCard>
+      <EvidencePackReviewSection evidencePacks={evidencePacks} />
 
       <SectionCard
         subtitle="Subordinate alert, evidence, and reconciliation context stays visible but secondary to the authoritative case state."
@@ -497,23 +398,4 @@ export function CaseDetailPageBody({
 
 function readBackendLinkedEvidencePacks(data: UnknownRecord) {
   return asRecordArray(getPath(data, "linked_evidence_packs"));
-}
-
-function evidencePackConflictTone(
-  conflictState: string | null,
-): ReturnType<typeof statusTone> {
-  if (conflictState === "conflicting") {
-    return "warning";
-  }
-  return statusTone(conflictState);
-}
-
-function evidencePackSourceTone(sourceState: string | null): ReturnType<typeof statusTone> {
-  if (sourceState === "unavailable") {
-    return "error";
-  }
-  if (sourceState === "degraded") {
-    return "warning";
-  }
-  return statusTone(sourceState);
 }
